Updated Click Handling and Started on Bottom Slots
New clickable interface to be applied to every clickable object. The click handler now just calls the clickable interface's onClick method, instead of spawning a ball on it's own, to allow for new types of clickable areas. Also, created the colliders for the bottom slots and added an empty script.