fixed ball getting stuck (again)
Setting friction to 0 helped alot, but the ball getting stuck still happened. I made the pin colliders into diamonds into spheres so less of a balance point, which helped, but it still happened. So, added a script to the ball to nudge it to either side if it sits still for too long. I think this will actually fix it completely.