PointHandler
Added PointHandler, a singleton script to handle all point calculations. Currently, it has the AddPoints() and FormatPoints() methods. AddPoints takes a basePoint value and adds it to the total, and FormatPoints returns a formatted string of the totalPoint value, which will be in scientific notation if it exceeds 1e4. AddPoints should be easy to expand when multipliers are added later.